POC: extend configschema.Attribute to include a NestedBlock
This POC adds an optional NestedBlock inside a configschema.Attribute.
The goal is to give provider developers fine-grained control over
structural attributes. From a users' perspective this should be
indistinguishable from an Attribute with a cty.Object type (something
which is not yet supported by the SDK), but it enables provider
developers to set flags (sensitive, computed, etc) on each nested
attribute.